+++ src/main/java/org/folio/search/service/setter/instance/CanceledLccnInstanceProcessor.java
@@ -1,14 +1,17 @@
-package org.folio.search.service.setter.authority;
+package org.folio.search.service.setter.instance;
 
 import java.util.List;
 import org.folio.search.integration.folio.ReferenceDataService;
 import org.folio.search.service.lccn.LccnNormalizer;
 import org.springframework.stereotype.Component;
 
+/**
+ * Instance identifier field processor, which normalizes LCCN numbers.
+ */
 @Component
-public class LccnAuthorityProcessor extends AbstractAuthorityLccnProcessor {
+public class CanceledLccnInstanceProcessor extends AbstractInstanceLccnProcessor {
 
-  private static final List<String> IDENTIFIER_NAMES = List.of("LCCN");
+  private static final List<String> IDENTIFIER_NAMES = List.of("Canceled LCCN");
 
   /**
    * Used by dependency injection.
@@ -16,7 +19,7 @@ public class LccnAuthorityProcessor extends AbstractAuthorityLccnProcessor {
    * @param referenceDataService {@link ReferenceDataService} bean
    * @param stringNormalizer     {@link LccnNormalizer} bean
    */
-  public LccnAuthorityProcessor(ReferenceDataService referenceDataService, LccnNormalizer stringNormalizer) {
+  public CanceledLccnInstanceProcessor(ReferenceDataService referenceDataService, LccnNormalizer stringNormalizer) {
     super(referenceDataService, stringNormalizer);
   }